+++ dimos/experimental/isolated_python/example/contract.py
@@ -18,11 +18,11 @@ from typing import Protocol
 
 from dimos.agents.annotation import skill
 from dimos.core.core import rpc
-from dimos.core.isolated_python_module import (
+from dimos.core.stream import In, Out
+from dimos.experimental.isolated_python.module import (
     IsolatedPythonModule,
     IsolatedPythonModuleConfig,
 )
-from dimos.core.stream import In, Out
 from dimos.msgs.std_msgs.Int32 import Int32
 from dimos.spec.utils import Spec
