--- examples/__init__.py
@@ -1,1 +1,0 @@
-"""Runnable DimOS examples."""
